requirements.txt
pandas
yfinance
numpy
matplotlib
PyPortfolioOpt
google-genai
streamlit
scikit-learn
plotly
